pyproject.toml
[project]
name = "strands-agents-qa"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "boto3>=1.38.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--tb=short",
    "--strict-markers",
    "--junitxml=test-results/results.xml",
    "--cov=age_calculator",
    "--cov=main",
    "--cov-report=xml:coverage.xml",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
]
markers = [
    "unit: Fast unit tests — no AWS credentials required",
    "integration: Integration tests with mocked AWS services",
    "evaluation: AI quality evaluation tests — dataset structural checks",
]

[tool.coverage.run]
source = ["age_calculator", "main"]
omit = ["*/tests/*", "*/.venv/*"]

[tool.coverage.report]
fail_under = 90
show_missing = true
